+++ examples/blog-win95/src/components/layout/index.tsx
@@ -1,15 +1,22 @@
-import { Footer } from "components/footer";
-import { FC, PropsWithChildren } from "react";
+import { PropsWithChildren } from "react";
 
-const Layout: FC<PropsWithChildren> = ({ children }) => {
+import { Footer } from "../footer";
+import { Header } from "../header";
+
+export const Layout: React.FC<PropsWithChildren> = ({ children }) => {
     return (
-        <div className="main">
-            <div className="layout">{children}</div>
-            <div>
-                <Footer />
+        <div>
+            <Header />
+            <div
+                style={{
+                    padding: "72px 24px 24px 24px",
+                    backgroundColor: "rgb(0, 128, 128)",
+                    minHeight: "calc(100vh - 48px - 72px - 24px)",
+                }}
+            >
+                {children}
             </div>
+            <Footer />
         </div>
     );
 };
-
-export default Layout;