+++ examples/blog-win95/src/index.tsx
@@ -1,10 +1,11 @@
 import React from "react";
-import App from "./App";
 import { createRoot } from "react-dom/client";
 
-const container = document.getElementById("root");
-// eslint-disable-next-line
-const root = createRoot(container!);
+import App from "./App";
+
+const container = document.getElementById("root") as HTMLElement;
+const root = createRoot(container);
+
 root.render(
     <React.StrictMode>
         <App />