+++ examples/blog-win95/src/pages/posts/index.ts
@@ -1,3 +1,0 @@
-export * from "./list";
-export * from "./edit";
-export * from "./create";