+++ packages/inferencer/src/inferencers/antd/show.tsx
@@ -199,6 +199,23 @@ export const renderer = ({
                 // ))}
                 // `;
             }
+
+            if (field.fieldable) {
+                return jsx`
+                <Title level={5}>${translatePrettyString({
+                    resource,
+                    field,
+                    i18n,
+                    noQuotes: true,
+                })}</Title>
+                <TextField value={${accessor(
+                    recordName,
+                    field.key,
+                    field.accessor,
+                )}} />
+                `;
+            }
+
             return jsx`
                 <Title level={5}>${translatePrettyString({
                     resource,