+++ packages/kbar/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @refinedev/kbar
 
+## 1.3.2
+
+### Patch Changes
+
+-   [#4948](https://github.com/refinedev/refine/pull/4948) [`8e5efffbb23`](https://github.com/refinedev/refine/commit/8e5efffbb231bc3163c56f8e823ccb649755a9d4) Thanks [@aliemir](https://github.com/aliemir)! - Keep the hook and component names in builds for better debugging.
+
 ## 1.3.1
 
 ### Patch Changes